requirements.txt
requests>=2.32.0
PyYAML>=6.0.1
pytest>=8.0.0
